--- a/tests/remote/providers.tf
+++ b/tests/remote/providers.tf
@@ -1,0 +1,7 @@
+provider "azurerm" {
+  features {
+    resource_group {
+      prevent_deletion_if_contains_resources = false
+    }
+  }
+}