--- a/tests/remote/resources.tf
+++ b/tests/remote/resources.tf
@@ -25,7 +25,11 @@ terraform {
 }
 
 provider "azurerm" {
-  features {}
+  features {
+    resource_group {
+      prevent_deletion_if_contains_resources = false
+    }
+  }
 }
 
 #