--- a/tests/remote/resources.tf
+++ b/tests/remote/resources.tf
@@ -25,12 +25,7 @@ terraform {
 }
 
 provider "azurerm" {
-  features {
-    key_vault {
-      purge_soft_delete_on_destroy    = true
-      recover_soft_deleted_key_vaults = true
-    }
-  }
+  features {}
 }
 
 #