This pull request resolves issue #19 by optimizing resource usage and test sequence, reducing deployment time by approximately from **~30 minutes** to **~10 minutes**.

1. **Reuse Existing Resources**  
   The module now utilizes existing resources such as the Recovery Service Vault, Key Vault, and Virtual Network, rather than creating new ones. This change reduces test deployment time by approximately 50%.

2. **Unique Resource Group Names**  
   A random string is now appended to the resource group name during remote test deployments. This allows for multiple concurrent test runs without conflicts.

3. **Single terraform deployment**  
   Previously, the remote test deployment involved three sequential test runs:

   - Creation of dependencies
   - Deployment of a Windows VM
   - Deployment of a Linux VM

   This pull request consolidates the process by performing a single terraform apply for both the dependencies and the VM deployments. This change further reduces test time by around 30%.

---

#### 👎🏻 Before – ⌛ 24:59

```console
$ time terraform test -test-directory=tests/remote
tests/remote/main.tftest.hcl... in progress
  run "setup_tests"... pass
  run "deploy_module_windows"... pass
  run "deploy_module_linux"... pass
tests/remote/main.tftest.hcl... tearing down
tests/remote/main.tftest.hcl... pass

Success! 3 passed, 0 failed.
terraform test -test-directory=tests/remote  26.25s user 9.62s system 2% cpu 24:58.81 total
```

#### 🥈 Use existing resources – ⌛ 13m25m

```console
$ time terraform test -test-directory=tests/remote
tests/remote/main.tftest.hcl... in progress
  run "setup_tests"... pass
  run "deploy_module_windows"... pass
  run "deploy_module_linux"... pass
tests/remote/main.tftest.hcl... tearing down
tests/remote/main.tftest.hcl... pass

Success! 3 passed, 0 failed.
terraform test -test-directory=tests/remote  71.18s user 29.27s system 12% cpu 13:25.20 total
```

#### 🥇 Use existing resources and do parallel deployment – ⌛ 9m30s

```console
$ time terraform test -test-directory=tests/remote
tests/remote/main.tftest.hcl... in progress
  run "deploy"... pass
tests/remote/main.tftest.hcl... tearing down
tests/remote/main.tftest.hcl... pass

Success! 1 passed, 0 failed.
terraform test -test-directory=tests/remote  13.66s user 4.57s system 3% cpu 9:30.04 total
```